+++ Asserv_finale/asserv/Utils/Utils.h
@@ -13,7 +13,7 @@ class Utils {
     // Conversions
     int64_t static mmToUO(Odometrie *odo, int64_t valeur);
     int64_t static degToUO(Odometrie *odo, int64_t valeur);
-    int64_t static radToUO(Odometrie *odo, int64_t valeur);
+    int64_t static radToUO(Odometrie *odo, double valeur);
     
 };
